added binary file

+++ Lecture02/hw2e.cpp
@@ -2,11 +2,13 @@
 // Created by Michael Moschetti on 9/11/17.
 //
 #include <iostream>
+#include <climits>
+#include <cstdint>
 using namespace std;
 
 //  MUST SUPPORT UP TO N = 30!
-unsigned long long fact(int x) {
-    unsigned long long factorial = 1;
+double long fact( double long x) {
+    double long factorial = 1;
     while (x > 1) {
         factorial *= x;
         x--;
@@ -14,7 +16,7 @@ unsigned long long fact(int x) {
     return factorial;
 }
 
-unsigned long long fact2(unsigned long long x) {
+double long fact2(double long x) {
     if (x == 1) {
         return 1;
     } else {
@@ -22,7 +24,7 @@ unsigned long long fact2(unsigned long long x) {
     }
 }
 
-int fibo(int f) {
+long int fibo(int f) {
     int result[f] = { 0 };
     result[1] = 1;
     result[2] = 1;
@@ -32,7 +34,7 @@ int fibo(int f) {
     return result[f];
 }
 
-int fibo2(int f) {
+long int fibo2(int f) {
     if (f <= 1) {
         return f;
     } else {
@@ -40,11 +42,13 @@ int fibo2(int f) {
     }
 }
 
-unsigned long long choose(unsigned long long n, unsigned long long r) {
+double long choose(double long n, double long r) {
     return fact2(n) / (fact2(r) * fact2(n - r));
 }
 
 int main() {
+
+    // cout automatically compresses to scientific notation
     cout << fact(5) << ' ' << fact2(5) << '\n';
     cout << fact(15) << ' ' << fact2(15) << '\n';
     cout << fibo(5) << ' ' << fibo2(5) << '\n';
